Move some reward utils to item/dialog utils.
